some fixes and updates (proposed by merlin)
git-svn-id: http://miranda.googlecode.com/svn/trunk@12307 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/Yahoo/im.cpp
@@ -133,7 +133,7 @@ void CYahooProto::ext_got_im(const char *me, const char *who, int protocol, cons
 
 		pre.timestamp = (DWORD)time(NULL);
 		
-		if (tm < pre.timestamp)
+		if ((DWORD)tm < pre.timestamp)
 			pre.timestamp = tm;
 		
 	} else
